refactor brainiac code
