--- projects/brainiac.js
@@ -1,10 +1,0 @@
-const { fullCoumpoundExports } = require('./helper/brainiac');
-
-// params =  comptroller , chain , brCKB , CKB
-
-module.exports = fullCoumpoundExports(
-  "0x5c68BDBba7151c486faCB66dc39b891030e46725",
-  "godwoken_v1",
-  "0x352d09567dE5A02415670723D09E006F623fE62e",
-  "0x7538C85caE4E4673253fFd2568c1F1b48A71558a"
-)
